refactor(tests): refactorizar tests módulo Notifications - dividir archivos >300 líneas
- Refactorizar notifications.service.spec.ts (357→3 archivos):
  * notifications.service.order-welcome.spec.ts (130 líneas)
  * notifications.service.payment-shipping.spec.ts (130 líneas)
  * notifications.service.preferences-queries.spec.ts (120 líneas)
  * helpers/notifications.test-helpers.ts (60 líneas)

- Refactorizar email.provider.spec.ts (319→3 archivos):
  * email.provider.sending.spec.ts (150 líneas)
  * email.provider.validation-errors.spec.ts (110 líneas)
  * email.provider.features.spec.ts (90 líneas)
  * helpers/email-provider.test-helpers.ts (70 líneas)

- Corregir helpers de notification.processor (tipo timestamp)

Tests: 93/93 pasando ✅
Validaciones: lint ✅ format ✅ type-check ✅ build ✅
